Some forgotten uint8_t's -> uint16_t's

+++ ports/raspberrypi/common-hal/analogio/AnalogIn.h
@@ -13,7 +13,7 @@
 typedef struct {
     mp_obj_base_t base;
     const mcu_pin_obj_t *pin;
-    uint8_t sample_size;
+    uint16_t sample_size;
 } analogio_analogin_obj_t;
 
 void analogin_init(void);
